third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGMarkerElement-dom-refX-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'refX' attribute of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.getAttribute('refX') is "500"
PASS markerElement.getAttribute('refX') is "5"
PASS successfullyParsed is true

TEST COMPLETE
